Replace commchannel
Replace ICCommLib channel used by Raidc since multiple guilds are using
it now, so everyone in one channel is not very nice. We'll now be using
RaidCore_<Partyleader>, and 5 secs after someone is promoted in the raid
we will re-join the channel of the new party leader. Note that this will
require everyone to update their clients to this version in order to
have syncs working.