add new events for boss to remove no break space char
There are some no break space inserted in boss names or datachron messages.
As they are difficult to see, the best is to replace them by space before
any processing.
So in next commits, boss will use this new event and older will be removed when
All boss will be changed.

Note:
 The GenericEvent name is a string, not a constant object.
 So I kept a similar name with carbine interface, by adding a prefix: 'RC_'.