Fix an issue that prevented OnReset() from working
The RAID_WIPE event was fired after the encounter was disabled, at that
point it doesn't work so OnReset() was never called. Since this is
required for some fights, I moved the event handler so it fires before
disabling the module.

This was an issue specifically on Avatus, where it only worked the first
pull, until another reloadui happened.